pyproject.toml
[tool.poetry]
name = "dimabot"
version = "0.1.0"
description = ""
authors = ["dima_st_bk <dimastbk@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
SQLAlchemy = {extras = ["mysql"], version = "^1.4.27"}
alembic = "^1.7.5"
python-dotenv = "^0.19.2"

gpxpy = "^1.5.0"
simplekml = "^1.3.6"
python-slugify = "^5.0.2"
geojson = "^2.5.0"

redis = "^4.1.3"
# pywikibot = {extras = ["mwparserfromhell"], version = "^6.6.3"}
# requests-oauthlib = "^1.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^21.10b0"
flake8-print = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
